WIP: split js
